@@ -1,27 +1,23 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-
-using NET.Core.Library.Domain.DBModels;
+﻿using NET.Core.Library.Domain.DBModels;
 using NET.Core.Library.Domain.Infrastructure;
 using NET.Core.Library.Domain.Infrastructure.EFCore;
 
-namespace ASP.NET.Core.WebAPI.Infrastructure.EFCore
-{
-    internal static class InMemoryDbDataSeeder
-    {
-        internal static void SeedTestData(IApplicationBuilder applicationBuilder)
-        {
-            var appDbContext = applicationBuilder.ApplicationServices.CreateScope()
-                                .ServiceProvider.GetRequiredService<AppDbContext>();
+namespace ASP.NET.Core.WebAPI.Infrastructure.EFCore;
 
-            if (appDbContext != null)
+internal static class InMemoryDbDataSeeder
+{
+    internal static void SeedTestData(IApplicationBuilder applicationBuilder)
+    {
+        var appDbContext = applicationBuilder.ApplicationServices.CreateScope()
+                            .ServiceProvider.GetRequiredService<AppDbContext>();
+
+        if (appDbContext != null)
+        {
+            foreach (Author author in TestData.GetAuthors())
             {
-                foreach (Author author in TestData.GetAuthors())
-                {
-                    appDbContext.Authors.Add(author);
-                }
-                appDbContext.SaveChanges();
+                appDbContext.Authors.Add(author);
             }
+            appDbContext.SaveChanges();
         }
     }
 }
